Stop using log_dir, move some reports to .cache.
Trying to reduce the amount of junk in project directories. The logs/
dir is now deprecated. dexy.log is now written to .cache and a reports/
directory exists in .cache, which reporters can use by setting a single
option. A deprecation message is printed the first time dexy encounters
an old logs/ directory.