FEATURE: Make it easier to create jinja2 filters.
jinja2 filters let you do things like `{{ foo | indent(4) }}`. A few
functions were exposed as filters but in a very ad-hoc, hard-coded way.
Now a `filters` setting controls which template plugins are exposed as
filters, and plugins can additionally specify that certain of their
elements be excluded (for example, the BeautifulSoup module is exposed
in the bs4 template plugin, but it doesn't make sense for this to be a
jinja2 filter). Existing filters were also cleaned up.